Refactor CheckoutBar to be reusable.
