+++ src/components/CheckoutBar/index.ts
@@ -1,0 +1,2 @@
+import CheckoutBar from './CheckoutBar';
+export default CheckoutBar;
